requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==3.3.0
Brotli==1.0.9
cairocffi==1.6.1
CairoSVG==2.7.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
cryptography==35.0.0
cssselect2==0.7.0
cycler==0.11.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap-v5==1.0.6
django-heroku==0.3.1
djongo==1.3.6
docutils==0.20.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.2
fonttools==4.42.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jwt==1.3.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.3.2
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.2
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.1
Panda3D==1.10.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psycopg2==2.9.1
pycparser==2.20
pydyf==0.7.0
Pygments==2.9.0
pymongo==3.12.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyphen==0.14.0
pypiwin32==223
pypng==0.0.21
PyQRCode==1.2.1
pyrsistent==0.17.3
pytesseract==0.3.10
python-dateutil==2.8.1
pytz==2021.3
pywin32==301
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
reportlab==4.0.4
requests==2.28.2
resolvelib==0.5.4
scikit-learn==1.0.1
scipy==1.7.0
segno==1.4.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
sqlparse==0.2.4
terminado==0.10.0
testpath==0.5.0
threadpoolctl==3.0.0
tinycss2==1.2.1
tornado==6.1
traitlets==5.0.5
translate==3.6.1
ttkbootstrap==0.5.1
typing_extensions==4.7.1
urllib3==1.26.15
virtualenv==20.9.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
zopfli==0.2.2
